requirements.txt
streamlit
audio-recorder-streamlit
google-generativeai
reportlab